requirements.txt
asgiref==3.2.10
Django==3.1.6
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
dj-database-url==0.5.0
django-import-export==2.5.0
django-easy-audit==1.3.1a1
hashids>=1.0.2
django-phonenumber-field>=5.2.0
phonenumbers>=8.12.26